5:50am. This might be my record for the earliest wakeup ever. Ergh, I am groggy. I'll start work on the language in a little bit.
I've been thinking about relative NG. It seems to me more like an unsupervised technique. It would in fact learn even if the gradients are zero. That is pretty amazing. And the x*x.T term with regards to inputs would stabilize it against blowing up. And since the actual part of the optimization would be in the last stage after all the gradients are propagated, that would have a great stabilizing effect for recurrent nets and RL tasks.

Yeah, I need this.

It actually seems like a technique that uses the gradients to do a kind of prediction on them.

It does remind me a bit of expectation maximization. I kind of want to do some reading on this.

https://arxiv.org/abs/1606.04838
Optimization Methods for Large-Scale Machine Learning

I saw some references for that in the above paper. I never knew variational inference was EM. But then again I do not know what VI is in the first place.

Let me do some reading on this until my brain comes back online and then I'll convert those hardcoded paths into compiler options.