Somehow I did not feel like reading where I left of yesterday so I ended up watching a lecture on variational inference.
It sort of ends up being policy gradient...

The natural gradient part of it only really concerns the update to single variables which makes it similar to relative gradient for single unit groupings.

Meh, the Bayesian stuff always boggles my mind. The Bayesian methods are going to turn into NNs by the end of all of this.

...

Forget that for a while. Right now what is on the agenda is to deal with the issue. Here I've factored one the options into a variable. Let me make sure everything still works and then I will turn them into parser macros and extend the record.

It is best to do it in two steps like this.

After I am done with the issue, I'll get back to reading. I fired a mail to Ke Sun asking him about whether he tried replacing the matrix inverse with SGD, so hopefully I should get a reply by tomorrow or in the next few days.

There is no point in starting work on RNG until that crucial bit is cleared out.

Now let me get this out of the way.