It is only 10am, but I've been reading papers for 4h now so let me take a break here. Mhhh. Despite getting up so early I slept well tonight unlike yesterday.
I read a bunch of stuff, but there was really no point. The RNG paper is as good as it will get.

It and CurveBall are the only two even remotely practical 2nd order method papers that I've come across. I hate overly mathematical papers like Olliviers'.

What I will do here is stop reading random stuff and walk myself through the math in the RNG paper. After that I will move to the implementation of it minus the matrix inversion stuff.

If I do not get a reply from Sun by tomorrow I'll try my luck with Nielsen who is the second author.

Let me chill a little for a bit.