Ugh. I got a reply from Ke, but I've realized just now that I've made a serious misunderstanding. It is not the matrix inversion that is the problem. Rather for natural gradient to work, it is necessary to average the FIM or the RFIM in this case over all the training samples.
This is a serious problem as I thought I could improve the algorithm by not storing them at all. I guess I'll be making such amateurish mistakes for a while.

So the RNG idea is mostly dead, unless I want to go for really sparse architectures where each neuron is restricted only to a small amount of inputs..

This is a viable option, but I was not intending it for it to be this complicated.

I need to think about this more.