Considering that in relative NG a single neuron would require storage squared in the number of connections and having to solve a linear system at every step (or every k steps), I'd really only resort to such a method in desperation. The optimization steps would be really complicated to implement as well for more complicated architectures.
As Ke Sun mentioned that they developed a new type of neuron whose paper they are going to release next week so I might as well wait for that to come out. Today is Sunday so I'll only have to a week at most.

Since he mentioned the neuron without any prompting from me it may be the case that they've distilled the insights from the RFIM research into it.

Since I've been reading papers since 6am, I am going to call it a day here.

It is pity that I could not grasp anything concrete in my research and that my course is dictated by chance events, but until something comes out, I think I will go back to work on the ML library.

One thing I will do tomorrow is test out the players using different random seeds and then get to work on implementing PPO.

Them having a new neuron sort of rules out work on differentiable plasticity for the time being. Probably the generic matrix kernel would be good for both, but when you are curious about something coming out, there is the urge to wait and see what it is before committing.