Ke Sun sent me an early draft the scalable relative NG paper so I am going to make implementing it and tuning it my next project. I hadn't expected that he had already thought of a solution to the scaling issues with RNG.
He said it would take two days to implement it in TF, but in Spiral it will take longer as it is not made to be good for experimentation, but making optimized implementations instead.

That having said maybe I should give implementing it in TF or PyTorch a try?

Well, let me put that aside for now. Before I begin work on this, I want to do some reading on spectral networks. Also I want to learn some Bayesian statistics while I am at it.

Also as I haven't done this already, I think I will look at the code for the RFIM paper. I want to see how they deal with updates to the entire layer.

Assuming this works, the end game for me would be to combine with newer RNN architectures like differentiable plasticity and then use to do RL. That is my own goal, but since Ke Sun is my benefactor in all of this I really should do something to make it easier for him to do his own work like the aforementioned TF implementation.

Well, for now let me do some studying. I should at least figure out what the inverse RFIM projection is.