"1:50pm. https://youtu.be/c69a_viMRQU?list=PLFDbGp5YzjqXQ4oE4w9GVWdiokWB9gEpm
Let me stop this here. I can't take it anymore.

I want to understand information geometry, but it is an unholy mix of statistics and differentiable geometry, both of which are over my league. That last time I did anything Bayesian which was the PGM course back in 2015 the experience was poor. I should have started with a more basic course like this one. It was a mistake I am going to have to fix.

2pm. I'll stop work on Spiral for a bit and get familiar with TF. I've unpacked Ke's code and will take a look at it shortly.

The goal will be this - just get the thing running on Mnist for now. I won't bother waiting for RNG to converge. After that I'll clean the code up and adapt the new scheme for it. To think I would be touching Python after all the work on Spiral. The truth is though, I am not that against it. Spiral itself did introduce me to wonders of dynamic typing, so I might as well take advantage of it.

I do think that frameworks like TF might be better for this kind of research.

When I am done with all of this, I will implement it in Spiral. Hopefully it will live up to my expectations.

Let me get cracking."