I got TF to install without a hitch.
Unfortunately the RNG code does not work because of version incompatibilities. That is about what I expected. I'd have been suprised if it actually worked.

I have no choice but to rebuild it every step by hand up to the point where I can begin work on the scalable version of RNG.

Well, I won't hurry with this. It will take a while.

Tomorrow my goal will be to go through the TF tutorials and get a network working on Mnist. I'll move forward from there.