...I tried Ke's code out as I had nothing better to do at this time.
Surprisingly it was quite easy to get it to work. That type error I ran into was the only one and it was trivial to fix. Despite that I'll be doing it in Spiral. I just can't get into this kind of code. I'll build generic and optimized implementations of all the kernels by hand in my own and favorite language where everything makes sense. That is the way to go. That settles it.