Instead of wasting time, I should have just started coding when I got up 1.5h ago. That Bayesian statistics course suddenly took a turn into boredom.
Somehow I am going to have to find a way of understanding information geometry without necessarily understanding everything about statistics.

...Ah, yeah, I forgot about that 50 page paper from 1998 which was on IG as applied to deep learning.

Well, forget it for a bit.

Right now, I've started this out by splitting the Cuda specific segments from the library. I will start by adding the special relative NG layers. In order to make this work, I am going to have to change the run function.

I'll remove the run_parallel for the time being. Also since I am just experimenting, the goal is first to get this up and running. Assuming I can get it to work well on the poker game and the deep autoencoder and maybe the char RNN, I will scale it up.

I'll have to implement softmax using an affine layer most likely. The standard feedforward ones are just too heavily interconnected when used together with RNG.

I'll derive it later.

First comes the design.

What is the problem?

Well, with RNG in the picture, the optimization pass does nested optimization of its own and takes in the inputs and outputs of the layer. It does not just take, but takes across timesteps.

I am going to have to use a list for that.

Furthermore, there are already some chinks in the backpropagation armor.

Even ignoring RNG for a moment, consider the case of batch norm.

The original paper had this weird thing where it gathers the statistics across the minibatch, but distilled to its core, batch norm is an unsupervised method that normalizes unitwise based on the expectation over the whole training set. It does not do so based on the gradients.

The way the library is structured at the moment, I have no way of implementing this.

I mean where would I put it. In the forward pass like the cuDNN? Pfft. In the backward pass? No. The backward pass is not intended to do any optimization, but merely to propagate gradients.

In order to implement batch norm, similarly to RNG, I would need to extend the current system.

There are further considerations. Consider EigenNets and decorrelated batch norm.

It is a decent assumption that I could get their benefit by replacing what they are doing with an autoencoder trained in an online fashion. For this in particular I'd want to keep its gradients separate from the supervised learning gradients. I'd only want it to do whitening.

Interestingly the inverse RFIM approximation that I am going to be doing in the optimization pass is also an unsupervised learning method as it depends only on the inputs and not on the actual gradients.

In the end only the network weight would depend on the SL gradients.

So batch norm, whitener autonecoder and RFIM approximators would not depend on the gradients. With 3/4 that really goes a long way in making NNs unsupervised and it would be interesting that most of the effort would go into making the optimization easier for the SL part.

I need to go forward with this scheme. I have no choice but to bring it all together.

Let me take a look at the paper by EigenNets guy and I'll start work on the library. Today my focus will be dealing on the design. After I get into the grove I'll implement the sigmoid NG layer.