https://arxiv.org/abs/1503.05671 Optimizing Neural Networks with Kronecker-factored Approximate Curvature
I skimmed this paper yesterday, and now it is time I properly go into it today. I'll give it my top focus. I need to understand all of this before I can come to a conclusion as to what makes the most sense with regards to natural gradient methods.

Ever since I started in April, my predictions have been extremely poor. It is no wonder. The reason why that is happening is because I am using simplistic heuristics to predict specific details of a highly technical field. Understanding the math behind all of this would get me a lot further.

When I am done with my studies, rather than trying to guess the best method, I'll start by implementing relative NG, and then compare them with the others.

Without a doubt, natural gradient will have a role to play in the future optimization and I need to grasp it before I can move on to the next step.

If I was doing SL I could go with the popular methods, but for RL higher standards are needed.