It is a failure. The KFAC paper is beyond my ability to grasp. Unlike yesterday when I internalized a good chunk of the old papers that I reread, here I got jack by focusing. Well, maybe I got a little, but it is nothing concrete.
Well, forget it that one for a while. It does not really matter.

I think that if I really focus, that I might be able to get the semi first order methods like RNG and PRONG. Maybe I'll replace the whitening transform the later does with matrix inverses using something else.

Let me take a break here. I've been at this for 3.5 hours.