"9:30am. I need to focus. Right now between reading papers and flying all over the place, my focus is scattered.
What are my goals here? Should I do this or study linear algebra as per plan?

I think I should change my plan to studying the 3 papers. If I figure out how to do whitening without the need for eigendecomposition that would be a big thing. Not to mention I would not need to understand eigendecomposition in that case.

Well, I should still do that, but how about I leave that for later?

It feels like all the pieces are falling into place. Being able to whiten the network would be enough to ensure the stability in the RL domain allowing me to return to work on the RL agents. I'll internalize this, implement and test out RNG and then move to implementing and testing out iterative whitening and use it to implement PRONG and BPRONG.

For now let me gather my focus and spend some good time studying the 3 papers. After that I'll study mirror descent and linear algebra."