The overall structure of the EASS paper is clear to me, but I need to be walked through the individual steps. I have the 824 page Handbook of Blind Source Separation sitting on my hard drive. I'll go through it tomorrow.
Right now as usual, I am just too tired.

After I am done with this, I'll study the mirror descent paper and linear algebra for a bit. There is also the ...

https://arxiv.org/abs/1511.06856
Data-dependent Initializations of Convolutional Neural Networks

...paper waiting for me.

Whitening and PRONG is really just the tip of the iceberg of unsupervised learning in NNs. The field should be collectively embarassed that I have to spend my time doing research instead of pillaging as I should.

I must admit, seeing what difference BPRONG and EigenNets make is really making me change my views on backprop. Without regularization it is pretty useless. The information geometry guys are the only ones who know what they are talking about.

I can't wait to start work on the implementations. This is going to be so great. Once I properly regularize the networks, the training for them should be much more robust and I will be able to plug in differentiable plasticity and such.

I am going to have to look into convex optimization to see if there are better alternatives in the optimization stage for the whitener than SGD.

Given that whitening is a form unsupervised learning, I am really wondering about one thing.

If I had a RNN that could predict the next inputs into a layer how would it be possible to take advantage of that? Decorrelating the individual layers is just a subset of the prediction task.

It should be possible to draw more power from this, but the reason and the theory of it all is missing. Until the light comes on, I and everybody else will have to keep on wandering in the dark.