Learning in manifolds - the case of source separation
I found the above paper on Genesis. It should explain to me the connection between natural and relative gradient learning.

Damn that handbook. It is not the kind of thing suited for casual reading.

Let me figure out what relative gradient is doing and then I will figure out what the whitening update is doing.