Let me finally resume.
I think at this point in time it is only sensible to despair a little over not being able to reason through optimization algorithms.

Damn that Learning on manifolds paper. I can't understand a thing of it.

I need to do more research on relative gradients. And I need to find a how the whitening matrix update is derived.