https://www.reddit.com/r/reinforcementlearning/comments/8pzl15/r_reinforcement_learning_hidden_theory_and_new/?st=ji8lca14&sh=05dd5207
https://youtu.be/dhEF5pfYmvc
Reinforcement Learning: Hidden Theory and New Super-Fast Algorithms (part 1)

https://youtu.be/Y3w8f1xIb6s
Reinforcement Learning: Hidden Theory and New Super-Fast Algorithms (part 2)

I have just finished watching the first part of this and am going to take a break. It is nice to watch this again and find that I can now grasp the structure of his arguments. Of course, I still do not get the details.

This is well worth studying though. It is a key piece of the puzzle to figuring out how to make bootstrapping RL work.

I'll have BPRONG to make semi-natural gradient updates in my NN, and I'll figure out how to throw this on top of that.

Well, I first have to understand it though, but I'll get it all in due time. Most of the developments in ML strike me as filler, but this is probably core. Along with natural gradients and new recurrent architectures and optimizers, I will go a long way once I put the pieces together.

But natural gradient updates and this will be the most essential. They are the foundation to ensuring the stability of all that will come after so I need to take care of it.

The time spent studying this will well pay off down the line.