That trimmed it down a little. Now I have the basics and none of the cruft. This is not the end of it, but it good enough to start things off.
Since I am not sure where to start work directly on the library, I'll sketch out a path by writing a test first. It won't be too hard. Ultimately, `rng` can be represented as a single sigmoid layer. There are other kinds of possible layer, but I will restrict myself to this. I only want to test the sigmoid rng layers on Mnist and autoencoder tasks and then I will move on to the next thing for the simple reason that RNG is hard to implement even for moderately harder things like softmax.