I am completely unfosed. Or more precisely, I am now thinking about how I am going to take the derivative through the log det and am thinking what am I going to do.
The trouble over doing this in Spiral rather than Tensorflow is that I would need to implement choleski decomposition and all that crap I am not interested in.

I've decided - if I cannot find a version of doing the derivative without computing the inverse, I will just move to doing BPRONG and apologize to Ke Sun later.