https://www.reddit.com/r/MachineLearning/comments/8qaqa1/d_how_would_projected_natural_gradient_work_with
"Trying to work out the [BPRONG](https://arxiv.org/abs/1704.07147) updates is no problem as long as they are on the linear regime. I can extend the equations readily for vanilla recurrent nets and probably could for convolutional nets. Likewise for nets with skip connections.

Here a short explanation of the problem.

    R (W U (z - c) + b) = R' (W' U' (z - c') + b'), for all z

`R` and `U`  are whitening matrices and `W` is the standard feedforward multiply. `z` is the input and `c` is the estimated mean of `z`. `W` and `b` are updated using SGD, and `R`,`U` and `c` are updated in an unsupervised fashion much like batch norm. After that last step, `W` and `b` need to be reprojected to `W'` and `b'` so that the function itself does not change. That is accomplished if the above equation is satisfied.

That can be done by setting `W' = R'^-1 R W U U'^-1` and `b' = R'^-1 (R W U c' - R W U c + R b)`. This is similar to how it is done in the paper, I've merely merged the forward and the backward updates. The method in the paper is not iterative and could be made a lot faster by updating the whitening matrix iteratively and doing the above updates without resorting to matrix inversion. The first could be done by taking the scheme for the update of the whitening matrix from the [EASS paper](https://www.semanticscholar.org/paper/Equivariant-adaptive-source-separation-Cardoso-Laheld/ca9fc6bb80d7fd89055f46d716aa8b6e2e377a5e) and the second using SGD or maybe some other convex optimization method.

How could the method be adapted to work when the update is something like from the [differentiable plasticity paper](https://arxiv.org/abs/1804.02464):

    x j (t) = o(sum (fun i -> w i j * x i (t-1) + a i j * Hebb i j (t) x i (t - 1)))
    Hebb i j (t+1) = n * x i (t - 1) * x j (t) + (1 - n) * Hebb i j (t)

The issue is how to make sure that the function does not change after updating the whitening matrices and the centering vectors. When the linear matrix multiply is replaced by a nonlinear function that is no longer obvious.

    R (f (W, U (z - c)) + b) = R' (f (W', U' (z - c')) + b') for all z

The above can no longer be solved by setting `W' = R'^-1 R W U U'^-1`. Can anything be done here to solve this?

It might be plausible that BPRONG could work without reprojecting the weights after the unsupervised updates if they are smooth enough which could be the case if they are done iteratively, but that does not feel right though."

Trying to find the right solution is so difficult.

I'll see if I get anything by tomorrow.

Also I've changed my mind on the RFIM again. I'll put in the matrix inverse tomorrow.

Really, the damn matrix inverse is all I need. Maybe the trace as well, but that will be trivial to implement.

Today I wanted to do all at once and overburdened myself. I'll try keeping a more even keel tomorrow. Right now I am too tired to do it. I'll just check out the API.