"7:45pm. Let me update the post.
"I think I get it now. I haven't gone all the way through, but unless I am mistaken differentiable plasticity could be written out as this in matrix form.

    Hebb (t+1) = n * x_in (t-1) * x_out (t) ^T + (1-n) * Hebb (t)
    x_out (t) = o(W * x_in (t-1) + (A .* Hebb (t)) * x_in (t-1))

`x_in (t-1) * x_out (t) ^T` is an outer product here. `.*` is a Hadamarad multiplication and the `*` is overloaded to be either scalar or matrix multiplication. Since the linear part is visible when written out like the above, there should be some way of sneaking the BPRONG updates in there."

8pm. It could even be written out like `x_out (t) = o((W + A .* Hebb (t)) * x_in (t-1))`

8:05pm. I cleared out this mystery. There is nothing linear algebra cannot do. Nothing.

That is a weight off my shoulder. My stress is really high right now.

I'll dedicate tomorrow to putting in the matrix inverse function then. I'll need it for the RNG. I'll also need the pointer handing thing for when I do skip connections. Honestly, passing arrays of pointers into kernels is what is my biggest concern right now. I'll need to experiment to make sure that it works right.

I won't be able to do that in a snap. It will take some debugging. But after that implementing the rest should not be that difficult."