"4:55pm. AAAAAAAAAAAAAhhhhhhhhhhhhhhhh....
I am such a moron. I am really such a moron.

Here is how you do it.

```
    Hebb (t+1) = n * W_u * U * (x_in (t-1) - c) * x_out (t) ^T + (1-n) * Hebb (t)
    x_out (t) = o(T * W_u * U * (x_in (t-1) - c) + (A .* Hebb (t)) * W_u * U * (x_in (t-1) - c))
```

And then back whitening could be added like this...

```
    Hebb (t+1) = n * W_u * U * (x_in (t-1) - c) * x_out (t) ^T + (1-n) * Hebb (t)
    x_out (t) = o(R * W_r * (W * W_u * U * (x_in (t-1) - c) + (A .* Hebb (t)) * W_u * U * (x_in (t-1) - c)))
```

Well, that is one mystery down. With this insight I could whiten anything. When written like this, I the forward and the backward whitening parts can be done in isolation. It just did not occur to me to rewrite it like this because I was locked in on W being a part of U. Whops.

Well, one extra tidbit is whether it would be possible to update both `W_u` and `U` simultaneously like in the EASS paper, but with centering added into the mix as well?

Oh actually, I do need the bias here as well.

So the real update would be something like...

```
    f x = W_u * U (x - c) + b
    r x = R * W_r * x
    Hebb (t+1) = n * f (x_in (t-1)) * x_out (t) ^T + (1-n) * Hebb (t)
    x_out (t) = o(r (W * f (x_in (t-1)) + (A .* Hebb (t)) * f (x_in (t-1))))
```"

As soon as I mailed the author and gave up on this, I figured it out. Wonderful. This is like the 100th time this happened to me.

Well, I really am happy that I went through this. Now I have a complete projected natural gradient method.

The reverse pass rather than doing it like the above could be done using a single matrix and updated using a single relative gradient update as shown in the EASS paper.

In fact I think the same could be done for the forward part as well. The bias would need to be done separately though, but that would be no problem.

Yeah, I see it now.

Since the EASS paper was from 96, it might be worth travling through the literature, but it is not really a priority at the moment. I'll put studying this for later.

Good.

With this I understand BPRONG completely.

I have the natural gradient method worth using. And with that I can finally begin programming.

I'll put in the matrix inverse tomorrow and then get to work on RNG according to the plan.