I focused for a bit. Here is the `matinv_batched`. There are 8 asserts in it and the phrase assert appears 11 times.
Damn what a pain. To be honest, I expected I would need a kernel to strip out the tensor pointers, but that seems to have been a mental error on my part.

Assuming it works this is quite nice. It is definitely pleasing to do a bit of work like this.

What I need to do next is the interface for it. It won't be hard at all. The tests will come after that. Thankfully I do not need to think about the column major vs row major issue since `(A^-1)^T = (A^T)^-1`.

Let me stop here so I can have breakfast first.