"For `c2`, let me look up the derivative of the `log det`.
11am. Uah, that `1 / o^2` term will be particularly troublesome. What the hell do I do about it? Well, I'll init it to one, but if this turns out to be particularly unstable I won't bother with fixing it too much and I'll make it a hyperparameter instead.

http://blog.as.uky.edu/sta707/wp-content/uploads/2014/01/matrix-derivatives.pdf

11:05am. Let me stop here. I am fully focused right now, but I should eat before I go any further.

Literally none of the parts of this are going to be quick or easy to implement. Before I can do the second term of the cost, I am going to have to put in batched matrix multiplication in.

The fourth term won't be a piece of cake either. I'll be very lucky if I manage to do it today."