I got cut off towards the end there, but I'll do some more today.
"I figured out the transpose thing. I did not get to ask. The logic is simple `a x b = (q x w) x b`. Transpose being `a x b -> b x a` implies that `b x a = b x (q x w)`. Because if it was `b x (w x q)` the right side would not be `a` and the transpose would go against its definition. Great."

Well, it was just as well that I stopped where I did.

It is usually the case that when the relatives come around with their kid in tow that I get some games for him that I end up playing myself.

I do not think I've played anything to speak of since he was here for a summer vacation last year and now I have Watch Dogs 2, Deus Ex HR, Max Payne 3 and will probably have something else by the end of next two weeks. It is just as well.

I really sacrificed quite a lot for the Spiral project, but I intend to get it all back.

The whole technology revolution needs to be brought into the next level.

Max Payne 1 and 2 are some of my favorite games of all time, but I want to halt this trend of constantly rehashing old games. It is unfair to MP3 which I am looking forward to playing, but the the true nature of technology is for it to be on the edge. When it becomes normal it looses its luster and turns into a commodity.

Maybe such a thing is inevitable, but not here and not now. Technology is aging a bit prematurely here - it is not old yet. It never even broke through its shell.

It never gave the people like me a chance to truly draw on its power. The debts must be repaid and I want power in exchange for my devotion.

Let me finish working out the math for the fourth term that I started yesterday, catch up to Keijo and then I will resume programming.