I am back. Honestly, I am still a bit unsure about the log det backwards step. I am going to have to make sure to do it properly when the time comes.
But forget about that. Damn, I want a break from this.

I knew that implementing this thing would be difficult and I see that my expectation has not been disappointed in the slightest.

Let me put it aside for now for a bit.

What I want to work on now is the new kind of init_redo kernel which will make it easier to make optimized reduction steps.

Once I am done with that, I'll do the proper forward and backward steps for each piece of the cost function.

This will take me a while.

Let me see if I can do one of the reduction kernels today.

The sketch that I've put into a text file is really reminiscent of how I used to program back in 2015. It reached its limit when I tried making a LSTM in this style and then could not get it to work.

After that I discovered AD and really fell in love with functional programming.

I can really do a lot better than this today.