I am thinking.
It is not that I can't do this.

It is just that the GPU experience in Spiral is ugly and unpleasant. Is this the best that I can do?

For a while now I've been having a vision of what GPU programming in Spiral really could be like.

First of all, I am going to make it so I do not have to compulsively write out to_dev_tensor everywhere. Honestly, I am so sick of that.

The second would be that I want to make it easier, much easier to write Cuda loops in Spiral.

It is nagging at me at the back of my mind, but I cannot quite draw the design out.

...

Ahhh, let me do some imaginary programming. Trully, if I had a Furthark-lite mini DSL in Spiral my life would be a lot easier.

I need to make it.

I've been working on these kernels for a while now. It is not like six months ago.

Spiral has a lot of very good things, but it not quite completely dominant yet.

I am sensing that there is a language that there should exist a language for GPU programming that should be the middle point between what I have in Spiral now and Futhark.

I should really thank Mr. Sun for setting me on this path. The cost function I've set out to implement is the definitely the most difficult NN related thing that I've yet to try implementing.

For a bit there I thought I would move on from language design, but it seems I am going to dedicate another bit of my life to it.

It is just as well. Making a programming language is the true essence of programming. It is what a real programmer draws power from and what sets apart them and the kiddies.

It is the connection to reality.

In order to master reinforcement learning I will need to master information geometry, but it will have to be done step by step.

There is no hurrying this.

With that in mind, let me make the idealized init_redo kernel.