Ah, fuck. I can't get anything out.
My mind is a lot more flid than the code I write, and I cannot solidify my ideas. It feels like the more I think about them the further they slip away.

So irritating, it feels like I wasted my whole day because of this. I had good momentum and it is gone now.

Well, I'll do what I can. I have an realistic idea of how I could do the init.

It would not be everything, but I could generalize the regular reduce kernels based on it.

GPU programming is just so difficult. It is the apex of programming in a sense. I really wonder how the Inspired would take care of it?