Well, I finished the init kernel, but I do not feel like doing anything else at this time.
It is annoying how messed up the library design is in general now. I haven't had the correct foresight with regards to the reduction kernels.

But actually despite everything, I am quite pleased with how the init kernel came out. I do not think I will be able to exceed this in design. The only thing I will do tomorrow is get rid of the compusive need to write to_dev_tensor everywhere. That will make the whole thing flow a lot more smoothly.

Given how it came out, the other init kernel is messed up now. The whole Cuda kernels library is incosistent in general.

I went to lengths to make it possible to reuse threads in it, but that was probably a waste of time.

Well, forget that. Tomorrow, my #1 thing will be to add the automatic conversions for device tensors.

After that I will look into adjusting the general design of it. The more generic I can make the library functions the better, but as today showed me that will be an uphil battle for sure.