It works now. I need to move on to the other kernel.
It occurs to me now that I've taken a break, that I've forgotten most of the API. That is never a good sign with regards to library design.

I've been thinking more in terms of init that anything else.

...Uah, now the next part is super annoying.

Right now I am thinking of how to avoid having to reverse the arguments in d2 redo and am coming up empty.

Since it will be a tranposed iteration that will create trouble if the user (most likely me) forgets that the arguments are coming in the opposite order.

Agh, let me just go for it.