Phew, added it.
Damn that was tedious.

For once, I think I am going to say to hell with it with consistency of kernels. I mean, I do not feel like rewriting literally everything once again just for this.

I am done with that sort of thing.

Here is where the rubber meets the road. Maybe I'll fix the kernels when I redo those parts of the library.

But now I need to laser on my goals.

I have the reduce kernels.

What is next?

Let me work on the cost once more.

I'll go through all the pieces again and do the forward and the back steps explicitly.