"3:50pm. https://www.youtube.com/watch?v=dza5JTvMpzk
Let me watch this.

4:05pm. Ok, I get it. But for at least CuBlas, I won't be able to reuse the LU matrix to speed up the inverse on the backward step. Well, unless I feel like adding forward and backward substitution myself. Agh...

Well, today was roughly as expected. Things like these always turn out to be huge time consumers. The is no simply in 'simply adding an API call'.

Let me go with eq 2 to start things off. It is merely `||g'^-1 g - I||`. Let me just say that it annoying that I do not have symmetric  tensors here, but regardless, with this particular approximation I won't have to worry about regularization terms for `g'^-1`. I just know that the `o` in `o^-2` will get pushed to zero at some point.

Let me sketch out the part for the L2 cost function."

Edit: Forgot to push the above.

At any rate, here is the completed optimization pass for a sigmoidal RFIM layer.

I'll work on actually getting it to run tomorrow and call it a day here. It is a nice bit work all in all.