So somehow yesterday I ended up playing Max Payne 3 for 5h and went to bed past 1am, so I am not my usual self right now. I did not mean for that to happen, but I just lost track of time. I definitely feel a bit fatigued right now due to missing some sleep.
Well, the game is pretty great. I think I preferred that combat of the previous games rather than the constant ducking behind cover of MP3, but it is fun nonetheless. The game is definitely a worthy successor. One thing that really pisses me off is how Max always switches to a handgun at every cutscene.

Well forget that.

It is time for some programming.

I had some time to think about where I want to put the optimization pass after I was done with work yesterday and I've decided to do it in the backward pass. Originally I wanted to add a whole separate pass, but that would be super complicated. The idea I have in mind right now is a lot better.