I am done with this. It is as I feared, RNG does not give any benefit at all as far as I can see, but has enormous computational costs. I cannot rule out that I made mistake somewhere, but I do not think this is the case here. My conclusion from this experiment with RNG is that projecting the weights directly is just a very hard task.
I only really started studying natural gradient methods 3.5 weeks ago, and for the first two weeks all the methods that I've seen used ineffciient inverses which they amortized by doing only every k steps where k is {1...1000}.

The RFIM paper by Ke Sun is unique from the rest the of NG related methods in that:

-It had suspiciously bad results. I thought by making it iterative this could be fixed.
-I understood it the first.
-I liked the story in the paper.

My reasoning going into this was not the greatest and a few days later I attained an understanding of the PRONG methods which at the time of writing this I believe to be objectively superior algorithms and can be made iterative much like relative NG.

That will be my next target.

I'll leave the code on the `rng` branch just in case I need it later.